[framework] Revamp pipeline creation.
